requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
python-magic==0.4.27

# LLM and AI
llama-index==0.12.30
llama-index-core==0.12.30
llama-index-llms-openai==0.3.33
llama-index-readers-file==0.4.7

# Utilities
loguru==0.7.3
python-dotenv==1.1.0

# AWS Lambda
mangum==0.19.0
aws-lambda-powertools>=3.0.0

# OCR
pillow==11.2.1
pytesseract==0.3.13
pdf2image==1.17.0

# New dependencies for caching and storage
psycopg2-binary==2.9.10
boto3==1.38.8